requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.1.0
charset-normalizer==3.1.0
discord==1.7.3
discord.py==1.7.3
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
motor==3.1.2
multidict==6.0.4
pymongo==4.3.3
python-dotenv==1.0.0
yarl==1.9.2